requirements.txt
Flask==2.2.5          # Core Flask framework
Flask-SQLAlchemy==3.0.5  # ORM for database interaction
Flask-JWT-Extended==4.4.4  # JWT-based authentication
Werkzeug==2.2.5       # Utilities and WSGI library (Flask dependency)
